Added a featured restaurants model.
ALso got all of the attributes into the restaurants dictionary. still need to move them into the db. looking at the proper way to do this. haven't got the "right solution yet but i should have it tomorrow"

I also need to tweak the importer to wait till it loops through all of the restaurants from the db before inserting a restaurant that hasn't been found.